requirements.txt
pytest
requests
allure-pytest
pytest-metadata